==> Lessons/06_Lab_GLMs_Isabel.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
  -the smaller the AIC, the better the model
  -good R2 values --> .7, .8, .9 
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 1:
Density <-
  ggplot(EPAair, aes(x = Ozone, fill = as.factor(Year))) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(values = c("darkgreen", "darkblue")) +
  labs(x = "Ozone AQI value", y = "Density", fill = "" )
print(Density)

Boxplot <-
  ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_boxplot() +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 4,
               label = c("b", "a")) + 
  labs(x = "", y = "Ozone AQI value") +
  ylim(0,150)
print(Boxplot)
  
  Violin <-
   ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_violin(draw_quantiles = 0.5) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 4,
               label = c("b", "a")) +
  labs(x = "", y = "Ozone AQI value") +
  ylim(0,150)
  print(Violin)
  
```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 
PM2.5_Ozone_regression <- lm(EPAair$PM2.5 ~ EPAair$Ozone)
summary(PM2.5_Ozone_regression)

#p-value is very small - which means we reject the new hypothesis (new hypothesis=there is no relationship)
#we don't want to analyze many variables in 1 model; simple is better!
#r-squared = .13 --> 13% of PM2.5 values are explained by Ozone levels


#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot
PM2.5_Ozone.Plot <-
  ggplot(EPAair, aes(x = Ozone, y = PM2.5)) +
  geom_point() +
  geom_smooth(method = "lm", col="blue")

print(PM2.5_Ozone.Plot)

```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.

#install.packages("corrplot")
library(corrplot)
EPA_subset <-
  EPAair %>%
  select(meanLat:Ozone) %>%
  na.omit() 

EPA_Corr <- cor(EPA_subset)
view(EPA_Corr) #Hard to visualize, can use "ellipse" to get a better visualization
corrplot(EPA_Corr, method = "ellipse")
corrplot.mixed(EPA_Corr, upper = "ellipse")

#The narrower the shape, the closer correlation
#Blue= + correlation, Red= - correlation
#Strong relationships between Ozone/PM2.5, PM2.5/meanLat, Ozone/Month

#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 4?

EPA_AIC <-
  lm(data = EPA_subset, PM2.5 ~ Ozone + Year + Month + meanLng + meanLat)
summary(EPA_AIC)
step(EPA_AIC)



#Overall p-value is very small, r-squared =  0.192
#Year has a value greater than .05 ...
#The more variables we add, the more complicated the model gets; AIC does the scale for us, "is it worth adding another variable to this model"
#We can make our results better by removing "Year", it will improve the AIC value


#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 5. 
OzoneBest.regression <-
    lm(data = EPA_subset, PM2.5 ~ Ozone + Month + meanLng + meanLat)
summary(OzoneBest.regression)
  
```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups
#Shows us the mean for each site
#Grouping the sites based on their mean dry mass

Litter.Totals <- Litter %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])
#Adding a column for "treatgroups"


# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot```

#Reorder sites based on "treatgroups" column (maybe in alphabetical order)
assign.group <- function(site)
  return

treatgroups=array(0,length(Litter.Totals))

Litter.Totals.plot2 <- ggplot(Litter.Totals, groups=treatgroups) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot2)
